requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.35
psycopg2-binary
python-jose==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
alembic
bcrypt==4.0.1
pywebpush
cryptography
